requirements/requirements.txt
--constraint "/usr/local/airflow/dags/constraints/constraints-3.11.txt"
--no-index
--find-links "/usr/local/airflow/plugins"

apache-airflow-providers-amazon==8.28.0
apache-airflow-providers-sftp==4.11.0
pandas==2.1.4
pendulum==3.0.0
boto3==1.35.7
psycopg2-binary==2.9.9
paramiko==3.4.1
tabulate==0.9.0
